rename table_check to ua_checker
